create mod_swap.py
1. `ag_swap.py` appended into `cg_swap.py`
2. `cg_swap.py` renamed to `mod_swap.py`
